--==real_engine==-- Well, this engine has become more than cmd modification I wanted before. I tried to make this many times but there always were flickering and lags. But now I have pretty good graphical instrument. My CEngine (this is new name for CmdEngine) handles special CFrameBuffer. This buffer has a console buffer handler, block of memory with CHAR_INFO and draws pixels in particular locations precisely cheking correctness of their position in the buffer.
Such happens that I store all objects in single memory chunk
which is allocated in the Init method of any of my engines.
And if there is overflow of framebuffer, I see the result in the moment
because it touches my RefKeepers and other objects which are allocated nearly.
And with that buffer I can update and clear the screen every frame.
Window configurations, pixel sizes, cursor and other configs are
on the side of CEngine which also has high-level methods for drawing.
Also, I've added states to this engine.
I can control engine behaviour via state class.

Also, there are added real widgets. Graphical user interfaces barely works!
Really, I have abstract CWidget class which gives derived CMenuItem.
We can switch states of CWidgets via enumerations and bind callbacks to them
For now I've made an example with button-items which changes draw color.
That buttons react on the pressing by mouse or ENTER key pressing.
They also change their color depending on the state we provide.
And the last thing - they can process events.
--==/real_engine==--